backend/requirements.txt
groq>=0.9.0
requests>=2.32.0
supabase>=2.4.0
python-dotenv>=1.0.0
tenacity>=8.3.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
google-generativeai>=0.8.0
